Re-enabled the bytecode offset option for the ASM5 backend
Soot dropped support for the `-keep-offset` flag with the switch to the ASM5
backend. This commit re-enables support for bytecode that does not use the JSRs.
The `JSRInlinerAdapter` class seems to rewrite the bytecode which would case the
offsets to be wrong. I changed `MethodBuilder` to instead extend `MethodNode`,
therefore any bytecode with a JSR will likely not work. Otherwise, all the
bytecode offsets are computed with the internal `MethodWriter` class, accessed
from `ClassWriter`. Currently I only add the `BytecodeOffsetTag` to
invokevirtual instructions, as that's all we needed for our purposes. It should
be possible to add tags to other instructions in a similar way.